% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{assembleTestRetest}
\alias{assembleTestRetest}
\title{Find test-retest data within a dataframe}
\usage{
assembleTestRetest(
  dataIn,
  qualitycolumns,
  subjectID,
  visitID,
  whichVisit,
  measureToRepeat,
  uniqueID
)
}
\arguments{
\item{dataIn}{Input data frame with the old name within}

\item{qualitycolumns}{the name(s) of quality-related column measurements}

\item{subjectID}{the unique subject id column name}

\item{visitID}{the column name that gives the visit / date; typically we look
for data the exists on the same visit.}

\item{whichVisit}{optional visit to use for the analysis}

\item{measureToRepeat}{the measurement to assemble for computing trt stats (e.g. ICC)}

\item{uniqueID}{optional column to contatenate to trt dataframe}
}
\value{
data frame with test-retest friendly organization
}
\description{
We assume there is a quality or related measure that will let us compute
distances between different time points and that these distances relate to
how similar two images are.  Note that the measurement may not relate to
quality at all but should map images into a similar metric space based on
appearance or some other objective quality that leads to the same value
when images are the same and changes continuously as images differ.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mytrt = assembleTestRetest( mydf, rbfnames[1], 'Id', 'visit' )
}
\author{
Avants BB
}
