% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgeRy.R
\name{averageNumpyArray}
\alias{averageNumpyArray}
\title{Average numpy array across batch}
\usage{
averageNumpyArray(numpyArray)
}
\arguments{
\item{numpyArray}{the array to average across rows (batch)}
}
\value{
the averaged arrays
}
\description{
Average numpy array across batch
}
\author{
Avants BB
}
