% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{biclusterMatrixFactorization}
\alias{biclusterMatrixFactorization}
\title{Matrix factorization biclustering}
\usage{
biclusterMatrixFactorization(mxdfin, measureColumns, k = 2, verbose = FALSE)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{measureColumns}{vector defining the data columns to be used for clustering.}

\item{k}{the number of clusters}

\item{verbose}{boolean}
}
\value{
a matrix with label identities; 0 indicates not in a cluster.
}
\description{
Cluster both rows and columns.  Return the joint cluster membership.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mybic = biclusterMatrixFactorization( mydf, rbfnames, k = 2 )
}
\author{
Avants BB
}
