% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgeRy.R
\name{chooseTrainingFilesToRead}
\alias{chooseTrainingFilesToRead}
\title{Choose files by modification date}
\usage{
chooseTrainingFilesToRead(fileDataFrame, random = FALSE, notFirst = FALSE)
}
\arguments{
\item{fileDataFrame}{a dataframe containing files required for training some model}

\item{random}{boolean that will return a random choice of files}

\item{notFirst}{boolean that will return a random choice of files instead
though not the most recently modified}
}
\value{
the recommended filenames to read right now
}
\description{
Choose files by modification date
}
\examples{
mydf = NULL
}
\author{
Avants BB
}
