% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{featureImportanceForSubtypes}
\alias{featureImportanceForSubtypes}
\title{subtype feature importance}
\usage{
featureImportanceForSubtypes(
  subtypeLabels,
  featureMatrix,
  associationType = c("features2subtypes", "subtypes2features"),
  visualize = FALSE
)
}
\arguments{
\item{subtypeLabels}{Input subtype assignments.}

\item{featureMatrix}{matrix/dataframe defining the data columns as features.}

\item{associationType}{either predictor features from subtypes or predict
subtypes from features.  will produce related but complementary results. in
some cases, depending on subtypes/degrees of freedom, only one will work.}

\item{visualize}{boolean}
}
\value{
dataframes for visualization that show feature to subtype importance e.g. via \code{pheatmap}
}
\description{
Associate predictors (or features) with subtypes; these could be diagnoses
or cluster assignments.  Will use regression to return data frames intended
to visualize the most important relationships between features and types.
There are two approaches - worth using both.  Can be combined with
boostrapping to give distributional visualizations.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
fimp = featureImportanceForSubtypes( mydf$DX, mydf[,rbfnames] )
fimp = featureImportanceForSubtypes( mydf$DX, mydf[,rbfnames], "subtypes2features" )
}
\author{
Avants BB
}
