% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{fillBaselineColumn}
\alias{fillBaselineColumn}
\title{fill baseline column}
\usage{
fillBaselineColumn(
  mxdfin,
  columnName,
  subjectID,
  visitID,
  baselineVisitValue,
  baselineExt = "_BL",
  deltaExt = "_delta"
)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{columnName}{string defining a valid column in the data frame.}

\item{subjectID}{the unique subject id column name}

\item{visitID}{names of the column that defines the variable defining
baseline-ness. e.g. \code{isBaseline}}

\item{baselineVisitValue}{the value defining baseline e.g. \code{TRUE}}

\item{baselineExt}{string appended to column name defining the baseline variable}

\item{deltaExt}{string appended to column name defining the change variable}
}
\value{
data frame with new columns
}
\description{
build a column of data that maps baseline values to every row. baseline values
starting points or reference points for each subject, e.g. a value of a measurement
at time zero.  the method will produce a mean value if multiple entries match
the subjectID and visitID conditions.
}
\examples{
mydf = generateSubtyperData( 100 )
mydf = fillBaselineColumn( mydf, "RandomBasisProjection01", "Id", "visit", "V0" )
}
\author{
Avants BB
}
