% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{filterByVisit}
\alias{filterByVisit}
\title{Get subjects within a given set of visits}
\usage{
filterByVisit(mxdfin, visitvar, visitselect)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{visitvar}{variable name for the visit column}

\item{visitselect}{the desired visits}
}
\value{
data frame
}
\description{
Get subjects within a given set of visits
}
\examples{
mydf = generateSubtyperData( 100 )
mydf = filterByVisit( mydf,  "visit", c("V0","V1"))
}
\author{
Avants BB
}
