% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{filterForGoodData}
\alias{filterForGoodData}
\title{Uses outlierness scoring to filter a data frame}
\usage{
filterForGoodData(
  dataIn,
  bestolvar = "OL_RKOF",
  quantileThreshold = 0.99,
  extraFiltering = FALSE,
  calck = 256
)
}
\arguments{
\item{dataIn}{Input data frame}

\item{bestolvar}{the outlier variable to prioritize}

\item{quantileThreshold}{the quantile to use for thresholing outlierness values}

\item{extraFiltering}{will perform brain-based application-specific filtering}

\item{calck}{optional parameter to pass to outlierness}
}
\value{
data frame
}
\description{
This will produce an inclusion function based on scores genetrated by the
\code{outlinerness} function within this package.  We assume the data frame
has values of the type \code{RandBasisProj*} or \code{RandomBasisProj*} within its column names.
}
\examples{
mydf = generateSubtyperData( 100 )
mydfol = filterForGoodData( mydf, calck = 8 )
}
\author{
Avants BB
}
