% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{generateDiskData}
\alias{generateDiskData}
\title{Generate augmentation data with on disk storage}
\usage{
generateDiskData(
  inputImageList,
  segmentationImageList,
  segmentationNumbers,
  selector,
  addCoordConv = 0,
  segmentationsArePoints = FALSE,
  maskIndex,
  numpynames,
  ...
)
}
\arguments{
\item{inputImageList}{list of lists of input images to warp.  The internal
list sets contains one or more images (per subject) which are
assumed to be mutually aligned.  The outer list contains
multiple subject lists which are randomly sampled to produce
output image list.}

\item{segmentationImageList}{of segmentation images corresponding to the input image list.}

\item{segmentationNumbers}{the integer list of values in the segmentation to model}

\item{selector}{subsets the inputImageList and segmentationImageList (eg to define train test splits)}

\item{addCoordConv}{boolean - generates another array with CoordConv data}

\item{segmentationsArePoints}{boolean - converts segmentations to points}

\item{maskIndex}{the entry within the list of lists that contains a mask}

\item{numpynames}{the names of the numpy on disk files should contain something with
the string mask if using maskIndex and something with the word coordconv if using CC}

\item{...}{additional arguments passed to the ANTsRNet function dataAugmentation}
}
\value{
list of array
}
\description{
Generate augmentation data with on disk storage
}
\examples{
mydf = NULL
}
\author{
Avants BB
}
