% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{hierarchicalSubtypePlots}
\alias{hierarchicalSubtypePlots}
\title{subtype plotting}
\usage{
hierarchicalSubtypePlots(
  inputDataFrame,
  variableToVisualize,
  hierarchyOfSubtypes,
  idvar,
  vizname,
  whiskervar = c("ci", "se"),
  consistentSubset = FALSE,
  manualColors,
  outputPrefix,
  width = 12,
  height = 8
)
}
\arguments{
\item{inputDataFrame}{Input complete data frame}

\item{variableToVisualize}{string naming the variable to display across subtypes}

\item{hierarchyOfSubtypes}{string vector of subtypes with increasing degrees of specificity}

\item{idvar}{variable name for unique subject identifier column}

\item{vizname}{the name of the grouped time variable (e.g. years change rounded to nearest quarter year)}

\item{whiskervar}{character either ci or se}

\item{consistentSubset}{display longitudinal data only from subjects that are consistently present at all visits}

\item{manualColors}{a list of user defined manual colors; the length of this list
should match the length of hierarchyOfSubtypes and colors should be named according
to the levels therein.  each entry in the list should be a string vector of color names.}

\item{outputPrefix}{filename prefix for the stored pdf plots; if missing, just plot to display}

\item{width}{the width of the graphics region in inches.}

\item{height}{the height of the graphics region in inches.}
}
\value{
the output is a set of plots saved at the outputPrefix location
}
\description{
Assemble a set of standard plots looking at subtype results to support comparing
across a hierarchy of types both cross-sectionally and longitudinally.
}
\examples{
mydf = generateSubtyperData( 1000 )
qdf = trainSubtypeUni( mydf, "cognition", c("C0","C1","C2"), c(0.33,0.66) )
qdf = predictSubtypeUni( mydf, qdf, "Id" )
\dontrun{
hierarchicalSubtypePlots( qdf, "cognition", c("DX", "subtype" ),
 "Id", "visit", outputPrefix='/tmp/X' )
}
}
\author{
Avants BB
}
