% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{highestQualityRepeat}
\alias{highestQualityRepeat}
\title{Get subjects and timepoints with the best quality}
\usage{
highestQualityRepeat(mxdfin, idvar, visitvar, qualityvar)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{idvar}{variable name for unique subject identifier column}

\item{visitvar}{variable name for the visit or date column}

\item{qualityvar}{variable name for the quality column; higher values should map to
higher quality data.}
}
\value{
data frame
}
\description{
This filters data that has both repeats (e.g. multiple images on the same day)
and (optionally) longitudinal data.  The quality measure should be a "higher"
is better" measurement.
}
\examples{
mydf = generateSubtyperData( 100 )
mydfhq = highestQualityRepeat( mydf, "Id", "visit", "quality")
}
\author{
Avants BB
}
