% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{loadNPData}
\alias{loadNPData}
\title{Read augmentation data from on disk storage}
\usage{
loadNPData(numpynames)
}
\arguments{
\item{numpynames}{the names of the numpy on disk files should contain something with
the string mask if using maskIndex and something with the word coordconv if using CC}
}
\value{
list of arrays in order of numpynames
}
\description{
Read augmentation data from on disk storage
}
\examples{
mydf = NULL
}
\author{
Avants BB
}
