% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\docType{data}
\name{nallspdma}
\alias{nallspdma}
\title{nallspdma}
\format{
A data frame (subset of columns described here):
\describe{
\item{SNP}{the SNP of interest}
\item{CHR}{associated chromosome}
\item{BP}{base pair}
\item{NearestGene}{nearest gene}
}
}
\usage{
data(nallspdma)
}
\description{
Supplementary table S2 from Nalls 2019 Parkinsons disease meta-analysis of
genome wide association studies.
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/31701892/}
}
\keyword{datasets}
