% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{outlierness}
\alias{outlierness}
\title{Outlierness scoring}
\usage{
outlierness(
  mxdfin,
  measureColumns,
  calck,
  outlierfunctions = c("LOOP", "LOF", "INFLO", "RDOS", "KDEOS", "LDF", "KNN_AGG",
    "KNN_IN", "KNN_SUM", "RKOF"),
  verbose = FALSE
)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{measureColumns}{vector string of column names to use in outlier detection}

\item{calck}{optional integer for knn}

\item{outlierfunctions}{vector of strings naming outlier functions to report}

\item{verbose}{boolean}
}
\value{
data frame
}
\description{
Produce several complementary measurements of outlierness
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mydf[8,rbfnames] = mydf[8,rbfnames] * 4.0
mydfol = outlierness( mydf, rbfnames )
olnames = names(mydf)[grep("OL_",names(mydf))]
mydf[6,olnames]
mydf[8,olnames]
}
\author{
Avants BB
}
