% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{plotSubtypeChange}
\alias{plotSubtypeChange}
\title{Plot longitudinal trajectories across subtypes}
\usage{
plotSubtypeChange(
  mxdfin,
  idvar,
  measurement,
  subtype,
  vizname = "timer",
  whiskervar = c("ci", "se"),
  extra = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{idvar}{variable name for unique subject identifier column}

\item{measurement}{the numeric variable to plot over time stratified by subtype}

\item{subtype}{the categorical (possibly ordinal) subtype variable}

\item{vizname}{the name of the grouped time variable (e.g. years change rounded to nearest quarter year)}

\item{whiskervar}{character either ci or se}

\item{extra}{additional naming variable}

\item{xlab}{additional naming variable}

\item{ylab}{additional naming variable}
}
\value{
ggplot
}
\description{
Uses \code{ggplot2} to visualize differences between subtypes over time.
}
\examples{
mydf = generateSubtyperData( 100 )
summ = plotSubtypeChange( mydf, "Id", "cognition", "DX", "visit" )
}
\author{
Avants BB
}
