% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{predictSubtypeClusterMulti}
\alias{predictSubtypeClusterMulti}
\title{Predict subtype from multivariate data}
\usage{
predictSubtypeClusterMulti(
  mxdfin,
  measureColumns,
  clusteringObject,
  clustername = "GMMClusters",
  idvar,
  visitName,
  baselineVisit
)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{measureColumns}{vector defining the data columns to be used for clustering.}

\item{clusteringObject}{a GMM object to predict clusters}

\item{clustername}{column name for the identified clusters}

\item{idvar}{variable name for unique subject identifier column}

\item{visitName}{the column name defining the visit variables}

\item{baselineVisit}{the string naming the baseline visit}
}
\value{
the clusters attached to the data frame
}
\description{
This is the inference module for subtype definition based on a matrix.
Currently only supports clustering based on gaussian mixtures
via the ClusterR package.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
gmmcl = trainSubtypeClusterMulti( mydf, rbfnames, maxk=4 )
gmmclp = predictSubtypeClusterMulti( mydf, rbfnames, gmmcl )
}
\author{
Avants BB
}
