% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{replaceName}
\alias{replaceName}
\title{Carefully replace a column name with a new one}
\usage{
replaceName(dataIn, oldName, newName)
}
\arguments{
\item{dataIn}{Input data frame with the old name within}

\item{oldName}{string column name to replace}

\item{newName}{the replacement name}
}
\value{
data frame
}
\description{
This will replace a column name with a new column name - it will not make
any changes to the underlying data.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mydf = replaceName( mydf, rbfnames[1], 'mileyCyrus' )
}
\author{
Avants BB
}
