% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgeRy.R
\name{specialCrop}
\alias{specialCrop}
\title{special cropping}
\usage{
specialCrop(x, pt, domainer = NULL)
}
\arguments{
\item{x}{input antsImage}

\item{pt}{point in physical space around which we crop}

\item{domainer}{vector of dimensionality equal to the image indicating the
size of cropped region}
}
\value{
cropped region
}
\description{
cropping method specialized for our data augmenation approach
}
\examples{
library( ANTsR )
specialCrop( ri(1), c(60,66), c(32,32) )
}
\author{
Avants BB
}
