% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{trainSubtypeClusterMulti}
\alias{trainSubtypeClusterMulti}
\title{Train subtype for multivariate data}
\usage{
trainSubtypeClusterMulti(
  mxdfin,
  measureColumns,
  method = "GMM",
  desiredk,
  maxk,
  groupVariable,
  group,
  frobNormThresh = 0.01,
  trainTestRatio = 0
)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{measureColumns}{vector defining the data columns to be used for clustering.
Note that these methods may be sensitive to scaling so the user may want to
scale columns accordingly.}

\item{method}{string GMM or kmeans}

\item{desiredk}{number of subtypes}

\item{maxk}{maximum number of subtypes}

\item{groupVariable}{names of the column that defines the group to use for training.}

\item{group}{string defining a subgroup on which to train}

\item{frobNormThresh}{fractional value less than 1 indicating the amount of change
in the reconstruction error (measured by frobenius norm) from the previous
iteration 1 - F_cur / F_prev that will determine the optimal number of clusters.
For GMM clustering.}

\item{trainTestRatio}{Training testing split for finding optimal number
of clusters. For GMM clustering.  If zero, then will not split data. Otherwise,
will compute reconstruction error in test data only.}
}
\value{
the clustering object
}
\description{
This is the training module for subtype definition based on a matrix.
Currently only supports clustering based on gaussian mixtures
via the ClusterR package.
One may want to use a specific sub-group for this, e.g. patients.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
gmmcl = trainSubtypeClusterMulti( mydf, rbfnames, maxk=4 )
}
\author{
Avants BB
}
