% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgeRy.R
\name{averageNumpyArraysFromDisk}
\alias{averageNumpyArraysFromDisk}
\title{Average numpy arrays from disk}
\usage{
averageNumpyArraysFromDisk(numpynames)
}
\arguments{
\item{numpynames}{the names of the numpy on disk files to average}
}
\value{
the averaged arrays
}
\description{
Average numpy arrays from disk
}
\author{
Avants BB
}
