% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{chooseTrainingFilesToRead}
\alias{chooseTrainingFilesToRead}
\title{Choose files by modification date}
\usage{
chooseTrainingFilesToRead(fileDataFrame)
}
\arguments{
\item{fileDataFrame}{a dataframe containing files required for training some model}
}
\value{
the recommended filenames to read right now
}
\description{
Choose files by modification date
}
\examples{
mydf = NULL
}
\author{
Avants BB
}
