% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{generateDiskData}
\alias{generateDiskData}
\title{Generate augmentation data with on disk storage}
\usage{
generateDiskData()
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{visitvar}{variable name for the visit column}

\item{visitselect}{the desired visits}
}
\value{
data frame
}
\description{
Generate augmentation data with on disk storage
}
\examples{
mydf = NULL
}
\author{
Avants BB
}
